+++ src/esercizio33/Rettangolo.java
@@ -3,7 +3,7 @@ package esercizio33;
 public class Rettangolo extends Forma{
 
     @Override
-    public int calcolaArea() {
+    public double calcolaArea() {
         return getBase() * getAltezza();
     }
 }
